+++ src/main/java/pages/BasePage.java
@@ -9,6 +9,7 @@ import org.openqa.selenium.support.ui.WebDriverWait;
 import utils.Driver;
 
 import java.time.Duration;
+import java.util.Set;
 
 public abstract class BasePage {