+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [1.0.4]
 
 ### Changed
-- [Assignment_18 Part 2 is completed]()
+- [Assignment_18 Part 2 is completed](https://inaracademygroup3.atlassian.net/browse/SCRUM-101?atlOrigin=eyJpIjoiY2VjMDcxYzY1NjUxNDY0ZjljYzYyYzYwZWZmNDA2NDEiLCJwIjoiaiJ9)
 - TestNG Framework Implementation
 - [Assignment_18 Part 1 is completed](https://inaracademygroup3.atlassian.net/browse/SCRUM-83?atlOrigin=eyJpIjoiZjU0ZjliZDVhNGVhNDg5NWE3NWJkMWJjMDE3ODM4NWQiLCJwIjoiaiJ9)
 - Web-automation test suite initialized